feat: refactor repository interfaces to improve namespace organization
